support both new and old testing system
While the migration to pytest is still under consideration, we should
not break existing CI. We can check for the existence of test.sh in
singularity and call it if it exists, otherwise do pytest.

Fixes: 59ffb09 ("update Containerfile to use the new python tests")

+++ start.sh
@@ -8,7 +8,12 @@ podman-compose build
 podman-compose up -d
 # wait until synapse is done initializing
 podman-compose logs -f submatrix 2>&1 | sed '/Synapse now listening on TCP port 8008/ q'
-virtualenv .
-pip install -r requirements.txt
-pytest
+if [ -f test.sh ]
+then
+	./test.sh
+else
+	virtualenv .
+	pip install -r requirements.txt
+	pytest
+fi
 podman-compose down
